% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruteforce.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Knapsack problem solver}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x:}{data frame of weights (first column) and value (second column)}

\item{W:}{Maximum capacity (numeric)}
}
\value{
Return the maximum value and elements (index of values)
}
\description{
take the weights, values as parameters and return maximum values and their possible combination
}
